[ch14865] Condition display of PSEA app content by environment flag
